Move a WeakActor to a new Zone:
A WeakActorName is unique inside a Zone and moving it keep this principle by throwing an exception.
A new WeakActorName can be proposed (and checked).

A WeakActor can only be in a group of its Zone.
When moving a WeakActor, you have to choose to have it removed from all groups, or throw.