Move a WeakActor while in Hierarchy with a workaround:
About the feature:
To know if the WeakActor is in any group outside of the destination Hierarchy, we compare the groups in the Hierarchy, with the groups of the WeakActor.
For the name clash, the new optional parameter of CK.fIsWeakActorNameInHierarchy is used (current ZoneId is excluded form the search).

About the workaround:
The TODO as been left as a remainder.
The TQL is not valid since there is a bug in the sql parser.
The SP is replaced whole.